Separate timer styling from app.css
- Create new folder for timer... src/newtab/timer
- Move timer.js to new folder
- Create timer.css and move styling from app.css
- Change import reference in index.js
