+++ src/newtab/timer/timer.css
@@ -1,0 +1,34 @@
+/* Timer Styling */
+
+.timer {
+  display: flex;
+  flex-flow: column;
+  font-family: sans-serif;
+  color: #fff;
+  font-weight: 100;
+  text-align: center;
+  font-size: 3rem;
+  margin: 20px;
+  width: inherit;
+}
+
+.timer .countdown {
+  display: flex;
+  flex-flow: row;
+}
+
+.timer .controls {
+  display: flex;
+  flex-flow: row;
+}
+
+.timer div > span {
+  padding: 20px;
+  border-radius: 30px;
+  background: rgba(0, 0, 0, 0.7);
+  display: inline-block;
+}
+
+.centiseconds {
+  color: #777;
+}
